Added authorization && fixed all the checks if user is logged in and if is own the post from the feed controller

+++ server/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -58,13 +59,19 @@ exports.createPost = (req, res, next) => {
     title,
     content,
     imageUrl,
-    creator: 'Alexandru', // TODO add user ID
+    creator: req.userId,
   });
-
-  // TODO add user as creator for post
 
   post
     .save()
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
     .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
@@ -138,7 +145,11 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
-      // TODO Add verification if user has permision to update post
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
 
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
@@ -176,11 +187,22 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
-      // TODO Check if user is logged in
+      // Check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
 
-      // TODO remove the image from storage
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
+    })
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then(result => {
       res.status(200).json({ message: 'Deleted post.' });
